chore: Move roundrobin package under connector
